# Conflicts:
#	pom.xml
#	src/main/java/swp/auctionkoi/configuration/SecurityConfig.java
#	src/main/java/swp/auctionkoi/controller/AuctionController.java
#	src/main/java/swp/auctionkoi/models/Auction.java
#	src/main/java/swp/auctionkoi/models/AuctionRequest.java
#	src/main/java/swp/auctionkoi/models/KoiFish.java
#	src/main/java/swp/auctionkoi/service/auction/AuctionService.java
#	src/main/java/swp/auctionkoi/service/auction/TraditionalAuctionService.java
#	src/main/java/swp/auctionkoi/service/auction/impl/AuctionServiceImpl.java
#	src/main/java/swp/auctionkoi/service/auctionrequest/impl/AuctionRequestServiceImpl.java

+++ src/main/java/swp/auctionkoi/dto/request/KoiFishDTO.java
@@ -18,7 +18,7 @@ public class KoiFishDTO {
     Integer breederId;
     String name;
     Sex sex;
-    Double size;
+    Float size;
     Integer age;
     String description;
     String image_Url;
